Time: 6 ms (49.08%) | Memory: 15.5 MB (56.73%) - LeetSync

+++ 122-best-time-to-buy-and-sell-stock-ii/best-time-to-buy-and-sell-stock-ii.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
-    int_fast32_t maxProfit(vector<int>& prices) {
-        int_fast32_t earn = 0;
-        for(short i=1;i<prices.size();i++){
+    int maxProfit(vector<int>& prices) {
+        int earn = 0;
+        for(int i=1;i<prices.size();i++){
             if(prices[i]>prices[i-1]) earn+=prices[i]-prices[i-1];
         }
         return earn;
